Add glossseealso template rule and support @linkend
I run into the problem of useing <glossseealso linkend="..."/>
instead of <glossseealso otherterm="..."/>.

As we are so get used to linkend in xref, I think it is
natural to support this in <glossseealso> too.

Although the commit contains several lines, the only change
in comparison to the original template rule is this:

  <xsl:variable name="otherterm" select="(@otherterm|@linkend)[1]"/>

The rest is like the original template rule.